refactor(cli): isolate macOS login helper installer
