+++ internal/cli/loginhelper/install_other_test.go
@@ -1,0 +1,20 @@
+//go:build !darwin
+
+package loginhelper_test
+
+import (
+	"context"
+	"testing"
+
+	"github.com/FlanChanXwO/pixiv-cli/internal/cli/loginhelper"
+)
+
+func TestInstallReportsUnsupportedPlatform(t *testing.T) {
+	cleanup, err := loginhelper.Install(context.Background(), "http://127.0.0.1/manual")
+	if cleanup != nil {
+		t.Fatal("unsupported platform returned a cleanup function")
+	}
+	if err == nil || err.Error() != "pixiv:// callback handler is only supported on macOS" {
+		t.Fatalf("unexpected error: %v", err)
+	}
+}
